卡片入库查询等增加卡面信息

+++ card/src/main/java/com/dili/card/dao/IStorageInDao.java
@@ -24,7 +24,7 @@ public interface IStorageInDao {
      * @param storageIn
      * @return
      */
-	int save(StorageInDo storageIn);
+	Long save(StorageInDo storageIn);
 
     /**
      * 根据id查询
